[refactor] TrainingEditModal -> EditTrainingModal 네이밍 변경
#40
