+++ src/components/EditTrainingModal/styled.ts
